Refactor project structure by removing unused files, updating package dependencies, and enhancing calendar components. Deleted index.html and popup.html, streamlined manifest.json, and improved cache handling in CourseRetrieval. Added CalendarResizeListener for better event display in the calendar view.
