--- index.html
@@ -1,12 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>MyPack Course Picker</title>
-</head>
-<body>
-  <div id="root"></div>
-  <script type="module" src="src/extension/popup.tsx"></script>
-</body>
-</html>
