--- public/popup.html
@@ -1,10 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <meta charset="UTF-8" />
-    <title>MyPack Extension</title>
-</head>
-<body>
-<div id="root"></div>
-</body>
-</html>
