+++ src/core/utils/dom.ts
@@ -64,7 +64,6 @@ export function waitForScheduleTable(): Promise<Element> {
         const observer = new MutationObserver(() => {
             const table = findTable();
             if (table) {
-                AppLogger.info(table);
                 AppLogger.info("Schedule table found by MutationObserver.");
                 observer.disconnect();
                 resolve(table);
