+++ src/course-management/components/DataGridCells/CalendarView.tsx
@@ -5,8 +5,8 @@ import {getCacheCategory} from '../../cache/CourseRetrieval';
 
 import {useEffect, useState} from 'react';
 
-import CreateCalender, {toMinutes} from './CalenderResizeListener';
-import type {ModifiedSection} from "../../../core/utils/CourseSearch/MergeDataUtil.ts";
+import CreateCalendar, { toMinutes } from './CalendarResizeListener';
+import type {  ModifiedSection } from '../../../core/utils/CourseSearch/MergeDataUtil';
 
 
 type ScheduleTableEntry = {
@@ -55,9 +55,10 @@ export interface ScheduleEvent {
 }
 
 
-export const CalendarView = (params: ModifiedSection) => {
-    // @ts-ignore
-    const {dayTime, courseData} = params;
+type CalendarViewProps = ModifiedSection & { dayTime?: string; courseData?: { code?: string } };
+
+export const CalendarView = (params: CalendarViewProps) => {
+    const { dayTime, courseData } = params;
 
     const [eventData, setEventData] = useState<ScheduleEvent[]>([]);
     const [isLoading, setIsLoading] = useState(true);
@@ -213,7 +214,7 @@ export const CalendarView = (params: ModifiedSection) => {
 
     return (
         <Box sx={{width: '100%', height: '100%'}}>
-            <CreateCalender eventData={eventData} />
+            <CreateCalendar eventData={eventData} />
         </Box>
     );
 };
