+++ src/course-management/services/api/DialogMenuSearch/searchService.ts
@@ -1,7 +1,6 @@
-import {AppLogger} from "../../../../core/utils/logger";
-import type {CourseData} from "../../../../core/utils/CourseSearch/ParseRegistrarUtil";
-import {formCourseURL, parseHTMLContent} from "../../../../core/utils/CourseSearch/ParseRegistrarUtil";
-import type {RequiredCourse} from "../../../../degree-planning/types/Plans";
+import { AppLogger } from '../../../../core/utils/logger';
+import { parseHTMLContent, formCourseURL } from '../../../../core/utils/CourseSearch/ParseRegistrarUtil';
+import type { CourseData } from '../../../../core/utils/CourseSearch/ParseRegistrarUtil';
 
 /**
  * Searches for open course sections based on term and course abbreviation/catalog number
@@ -105,13 +104,3 @@ export async function batchSearchOpenCourses(
     
     return null;
 }
-/**
- * This fucntion needs to go
- * @deprecated
- * @param term The academic term to search in
- * @param course The course requirement details
- * @returns CourseData or null if not found or error
- */
-export async function searchOpenCourses(term: string, course: RequiredCourse): Promise<CourseData | null> {
-    return searchOpenCoursesByParams(term, course.course_abr, course.catalog_num);
-}
