+++ src/course-management/services/api/PackPlannerAPI/courseDetail/courseDetailService.ts
@@ -82,7 +82,7 @@ async function fetchSingleCourseData(course: Course): Promise<SingleCourseDataRe
     
     // Generate hash and store in persistent cache
     const courseHash = await generateCacheKey(course.abr + " " + course.instructor);
-    await setGenericCache("courseList", {hashKey : courseHash,cacheData : JSON.stringify(combinedData)});
+    await setGenericCache("courseList", {[courseHash]:JSON.stringify(combinedData)});
     
     return combinedData;
   } catch (error) {
