+++ src/extension/content.tsx
@@ -10,16 +10,12 @@ import {setupListener} from "../degree-planning/services/siteResponseStorage.ts"
 import createCache from "@emotion/cache";
 
 
-
 declare global {
     interface Window {
         __savedDefine?: any;
         __mypackEnhancerInitialized?: boolean;
     }
 }
-
-const __psDefine = window.__savedDefine;      // pull from global
-delete window.__savedDefine;
 
 AppLogger.info("MyPack Enhancer script started.");
 
@@ -39,6 +35,7 @@ export function createEmotionCache() {
 }
 
 export const myEmotionCache = createEmotionCache();
+
 function debounceScraper(scrapePlanner: (plannerTableElement: Element) => Promise<void>) {
     // Debounce time of 100ms
     return debounce(async () => {
@@ -53,37 +50,6 @@ function debounceScraper(scrapePlanner: (plannerTableElement: Element) => Promis
 }
 
 const debouncedScrapePlanner = debounceScraper(scrapePlanner);
-
-/**
- * Enhanced message sending with retry logic for service worker restarts
- */
-async function sendMessageWithRetry(message: any, maxRetries = 3): Promise<any> {
-  for (let i = 0; i < maxRetries; i++) {
-    try {
-      const response = await chrome.runtime.sendMessage(message);
-      
-      // Check if response indicates success
-      if (response && response.success !== false) {
-        return response;
-      }
-      
-      // If we get a failed response, don't retry
-      if (response && response.success === false) {
-        throw new Error(response.error || 'Request failed');
-      }
-      
-    } catch (error) {
-      AppLogger.warn(`[Content] Message attempt ${i + 1} failed:`, error);
-      
-      if (i === maxRetries - 1) {
-        throw error;
-      }
-      
-      // Wait before retry with exponential backoff
-      await new Promise(resolve => setTimeout(resolve, 1000 * Math.pow(2, i)));
-    }
-  }
-}
 
 function injectXHRHookScript() {
     const script = document.createElement("script");
@@ -181,7 +147,7 @@ document.addEventListener('DOMContentLoaded', () => {
 (async () => {
     try {
         if (window.__mypackEnhancerInitialized) {
-            console.warn("MyPack Enhancer already initialized. Skipping...");
+            AppLogger.warn("MyPack Enhancer already initialized. Skipping...");
             return;
         }
         window.__mypackEnhancerInitialized = true;
@@ -239,12 +205,3 @@ document.addEventListener('DOMContentLoaded', () => {
     }
 })();
 
-if (__psDefine) {
-    Object.defineProperty(window, 'define', {   // hand the loader back
-        value: __psDefine,
-        configurable: true
-    });
-}
-
-// Export enhanced message sending for other modules
-(window as any).sendMessageWithRetry = sendMessageWithRetry;
