+++ src/extension/prelude.js
@@ -1,7 +1,1 @@
-﻿const __psDefine = window.define;
-Object.defineProperty(window, 'define', {
-    value: undefined,
-    configurable: true
-});
-// expose for later:
-window.__psDefine = __psDefine;
+﻿// legacy AMD interop removed; file no longer needed
