+++ src/ui-system/themes/AppTheme.tsx
@@ -1,14 +1,14 @@
 import * as React from 'react';
-import type {Components, Theme, ThemeOptions} from '@mui/material/styles';
-import {createTheme, Experimental_CssVarsProvider as CssVarsProvider} from '@mui/material/styles';
-import {inputsCustomizations} from './customizations/inputs';
-import {dataDisplayCustomizations} from './customizations/dataDisplay';
-import {autocompleteCustomizations, dataGridCustomizations} from './customizations/dataGrid';
-import {feedbackCustomizations} from './customizations/feedback';
-import {navigationCustomizations} from './customizations/navigation';
-import {surfacesCustomizations} from './customizations/surfaces';
-import {colorSchemes, shadows, shape, typography} from './themePrimitives';
-import {AppLogger} from "../../core/utils/logger.ts";
+import type { Components, Theme, ThemeOptions } from '@mui/material/styles';
+import { createTheme, Experimental_CssVarsProvider as CssVarsProvider } from '@mui/material/styles';
+import { inputsCustomizations } from './customizations/inputs';
+import { dataDisplayCustomizations } from './customizations/dataDisplay';
+import { autocompleteCustomizations, dataGridCustomizations } from './customizations/dataGrid';
+import { feedbackCustomizations } from './customizations/feedback';
+import { navigationCustomizations } from './customizations/navigation';
+import { surfacesCustomizations } from './customizations/surfaces';
+import { colorSchemes, shadows, shape, typography } from './themePrimitives';
+import { AppLogger } from '../../core/utils/logger.ts';
 
 interface AppThemeProps {
   children: React.ReactNode;
@@ -27,7 +27,8 @@ export default function AppTheme(props: AppThemeProps) {
     return disableCustomTheme
       ? {}
       : createTheme({
-          colorSchemes, // Recently added in v6 for building light & dark mode app, see https://mui.com/material-ui/customization/palette/#color-schemes
+          // Use the remote colorSchemes definition (supports CSS vars provider)
+          colorSchemes,
           typography,
           shadows,
           shape,
